BehaviorTree updates
- Make Child<S> private
- Added share to `reset` in `Action`
